- Обновленный генератор ассоримента торговцев - Заморозка интерфейса на время изменения размера окна - Доработка ProbabilityChecker - Смена значений FloatRangeParameter на double - Убран прямой доступ к словарю параметров в LtxSection

+++ Simple randomizer SoC/Forms/Dialogs/OrderableParameterListDialog.cs
@@ -230,7 +230,7 @@ namespace Simple_randomizer_SoC.Forms.Dialogs
                 minInput.Value = (decimal)parameter.MinValue;
                 minInput.ValueChanged += (s, e) =>
                 {
-                    parameter.MinValue = (float)minInput.Value;
+                    parameter.MinValue = (double)minInput.Value;
                 };
 
                 var maxInput = new NonScrollNumeric();
@@ -241,7 +241,7 @@ namespace Simple_randomizer_SoC.Forms.Dialogs
                 maxInput.Value = (decimal)parameter.MaxValue;
                 maxInput.ValueChanged += (s, e) =>
                 {
-                    parameter.MaxValue = (float)maxInput.Value;
+                    parameter.MaxValue = (double)maxInput.Value;
                 };
 
                 var precisionInput = new NonScrollNumeric();
